Add collect() method to regular collection
